requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.3.4
python-multipart==0.0.12
google-cloud-vision==3.7.4
pillow==12.0.0
slowapi==0.1.9
tenacity==9.0.0
structlog==24.4.0
bleach==6.1.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
httpx==0.27.0  
black==24.4.2
isort==5.13.2
pre-commit==3.7.1